requirements.txt
python-telegram-bot==20.7
python-binance==1.0.19
requests==2.31.0
python-dotenv==1.0.0
pydantic>=2.10.0
aiohttp>=3.9.1
typing-extensions>=4.8.0
openai>=1.0.0
google-generativeai>=0.3.0
